third_party/WebKit/LayoutTests/fast/dom/Window/window-scaled-viewport-properties-expected.txt

This test makes sure the window properties related to the viewport remain correct under pinch-to-zoom.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


===Initial Scale===

PASS window.innerWidth is 1600
PASS window.innerHeight is 1200
PASS internals.visualViewportWidth() is 1600
PASS internals.visualViewportHeight() is 1200

===Pinch Zoom in to 2X===

PASS window.innerWidth is 1600
PASS window.innerHeight is 1200
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS internals.visualViewportWidth() is 400
PASS internals.visualViewportHeight() is 300
===ScrollBy===
PASS window.scrollX is 100
PASS window.scrollY is 200
PASS internals.visualViewportScrollX() is 100
PASS internals.visualViewportScrollY() is 200
PASS window.scrollX is 400
PASS window.scrollY is 300
PASS internals.visualViewportScrollX() is 400
PASS internals.visualViewportScrollY() is 300
PASS window.scrollX is 0
PASS window.scrollY is 0
===ScrollTo===
PASS window.scrollX is 100
PASS window.scrollY is 200
PASS internals.visualViewportScrollX() is 100
PASS internals.visualViewportScrollY() is 200
PASS window.scrollX is 400
PASS window.scrollY is 300
PASS internals.visualViewportScrollX() is 400
PASS internals.visualViewportScrollY() is 300
PASS window.scrollX is 0
PASS window.scrollY is 0

===Pinch Out to 0.5X===

PASS window.innerWidth is 1600
PASS window.innerHeight is 1200
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS internals.visualViewportWidth() is 1600
PASS internals.visualViewportHeight() is 1200
===ScrollBy===
PASS window.scrollX is 100
PASS window.scrollY is 200
PASS internals.visualViewportScrollX() is 100
PASS internals.visualViewportScrollY() is 200
PASS window.scrollX is 400
PASS window.scrollY is 300
PASS internals.visualViewportScrollX() is 400
PASS internals.visualViewportScrollY() is 300
PASS window.scrollX is 0
PASS window.scrollY is 0
===ScrollTo===
PASS window.scrollX is 100
PASS window.scrollY is 200
PASS internals.visualViewportScrollX() is 100
PASS internals.visualViewportScrollY() is 200
PASS window.scrollX is 400
PASS window.scrollY is 300
PASS internals.visualViewportScrollX() is 400
PASS internals.visualViewportScrollY() is 300
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS successfullyParsed is true

TEST COMPLETE
